backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.20,<1
pydantic-settings>=2.6,<3
qdrant-client>=1.12,<2
sentence-transformers>=3.4,<4
openai>=1.55,<2
PyMuPDF>=1.24,<2
python-docx>=1.1,<2
orjson>=3.10,<4
pytest>=8.3,<9
httpx>=0.28,<1
